+++ src/gods/hermes.lua
@@ -75,7 +75,7 @@ gods.CreateBoon({
 gods.CreateBoon({
 	pluginGUID = _PLUGIN.guid,
     characterName = "Hermes",
-	internalBoonName = "OopsAllCursedBoon",
+	internalBoonName = "PurgeRarifyBoon",
     isLegendary = false,
 	InheritFrom = {
 		"SynergyTrait",
@@ -83,20 +83,20 @@ gods.CreateBoon({
     addToExistingGod = { boonPosition = 15 },
 	reuseBaseIcons = true,
 
-    displayName = "Royal Decree",
-    description = "Inflicting {$Keywords.Link} on foes applies every {$Keywords.Status} you can inflict using other abilities.",
-	StatLines = { "CursePotencyDisplay1" },
+    displayName = "Sacrosanct Alms",
+    description = "You can purge most rewards, and when you do, a random {$Keywords.GodBoon} you have gains {$Keywords.Rarity}.",
+	StatLines = { "RarityAmountDisplay1" },
     customStatLine = {
-        Id = "CursePotencyDisplay1",
-        displayName = "{!Icons.Bullet}{#PropertyFormat}Bonus {$Keywords.Link} Damage:",
-        description = "{#UpgradeFormat}{$TooltipData.StatDisplay1}",
+        Id = "RarityAmountDisplay1",
+        displayName = "{!Icons.Bullet}{#PropertyFormat}Rarity Increase:",
+        description = "{#UpgradeFormat}+{$TooltipData.ExtractData.RarityAmount}",
     },
 	requirements =
 	{
 		OneFromEachSet =
 		{
-			{ "HeraWeaponBoon", "HeraSpecialBoon", "HeraCastBoon", "HeraSprintBoon" },
-			{ "HermesWeaponBoon", "HermesSpecialBoon", "TimedKillBuffBoon" },
+			{ "DamageShareRetaliateBoon", "CommonGlobalDamageBoon", "BoonDecayBoon", "ElementalRarityUpgradeBoon" },
+			{ "MoneyMultiplierBoon", "RestockBoon", "ElementalUnifiedBoon" },
 		},
 	},
     boonIconPath = "Wistiti-HermesDuosBoonIcons\\HeraHermesDuo",
@@ -105,230 +105,19 @@ gods.CreateBoon({
 	ExtractValues =
 	{
 		{
-			Key = "DamageShareAmountIncrease",
-			ExtractAs = "TooltipAmount",
-			Format = "Percent",
-		},
-		{
-			ExtractAs = "DamageShareDuration",
+			Key = "RarityUpgrade",
+			ExtractAs = "RarityAmount",
 			SkipAutoExtract = true,
-			External = true,
-			BaseType = "EffectData",
-			BaseName = "DamageShareEffect",
-			BaseProperty = "Duration",
-		},
-		{
-			ExtractAs = "DamageShareAmount",
-			SkipAutoExtract = true,
-			External = true,
-			BaseType = "EffectData",
-			BaseName = "DamageShareEffect",
-			BaseProperty = "Amount",
-			Format = "Percent",
 		},
 	},
 
 	ExtraFields = 
 	{
-		DamageShareAmountIncrease = { BaseValue = 0.2 },
-		SetupFunction = 
-		{
-			Name = "BuildValidEffects",
-			Args = 
-			{
-				StatusTraitNames = 
-				{
-					DamageEchoEffect = LinkedTraitData.ZeusEchoTraits,
-					BurnEffect = LinkedTraitData.HestiaBurnTraits,
-					ChillEffect = LinkedTraitData.DemeterRootTraits,
-					LegacyChillEffect = { "DemeterSprintBoon", "CastNovaBoon", "StormSpawnBoon" },
-					AmplifyKnockbackEffect = { "PoseidonStatusBoon", "PoseidonCastBoon" },
-					WeakEffect = LinkedTraitData.AphroditeWeakTraits,
-					BlindEffect = LinkedTraitData.ApolloBlindTraits,
-					DelayedKnockbackEffect = { "MassiveKnockupBoon" },
-					AresStatus = LinkedTraitData.AresRendTraits,
-				},
-				ConditionalOutgoingDamageMultipliers = 
-				{
-					{
-						EffectName = "AresStatus",
-						OutgoingDamageModifiers = 
-						{
-							ValidProjectilesLookup = { 
-								"HeraCastDamageProjectile",
-								"HeraSprintProjectile",
-							},
-							MissingEffectDamage = EffectData.AresStatus.BonusBaseDamageOnInflict,
-							MissingEffectName = "AresStatus",
-							MissingDamagePresentation = 
-							{
-								TextStartColor = Color.AresDamageLight,
-								TextColor = Color.AresDamage,
-								FunctionName = "AresRendApplyPresentation",
-								HitSimSlowParametersFalseTraitName = "StaffRaiseDeadAspect",
-								SimSlowDistanceThreshold = 180,
-								HitSimSlowCooldown = 0.8,
-								HitSimSlowParameters =
-								{
-									{ ScreenPreWait = 0.02, Fraction = 0.13, LerpTime = 0 },
-									{ ScreenPreWait = 0.10, Fraction = 1.0, LerpTime = 0.05 },
-								},
-							},
-						}
-					}
-				}
-			},
-		},
-		OnEffectApplyFunction = 
-		{
-			FunctionName = _PLUGIN.guid .. "." .. "HitchCopyStatus",
-			FunctionArgs = 
-			{
-				ValidStatusNames = 
-				{
-					DamageEchoEffect = true,
-					BurnEffect = "ApplyBurn",
-					ChillEffect = "ApplyRoot",
-					LegacyChillEffect = true,
-					AmplifyKnockbackEffect = true,
-					WeakEffect = "ApplyAphroditeVulnerability",
-					BlindEffect = true,
-					DelayedKnockbackEffect = true,
-					-- AresStatus = true,
-				},
-			},
-		},
+		SpeakerNames = { "Hera" },
+		RarityUpgrade = 1, --description only
+		Permanent = true,
     },
 })
-
---[[gods.CreateBoon({
-	pluginGUID = _PLUGIN.guid,
-    characterName = "Hermes",
-	internalBoonName = "RarifyGoldBoon",
-    isLegendary = false,
-	InheritFrom = {
-		"SynergyTrait",
-	},
-    addToExistingGod = { boonPosition = 15 },
-	reuseBaseIcons = true,
-
-    displayName = "Corrputed Monarchy",
-    description = "You may spend {!Icons.Currency} to {$Keywords.RarityUpgrade} any {$Keywords.GodBoonPlural} offered to you.",
-	StatLines = { "CursePotencyDisplay1" },
-    customStatLine = {
-        Id = "CursePotencyDisplay1",
-        displayName = "{!Icons.Bullet}{#PropertyFormat}Rarify Cost:",
-        description = "{#UpgradeFormat}{$TooltipData.StatDisplay1}",
-    },
-	requirements =
-	{
-		OneFromEachSet =
-		{
-			{ "HeraWeaponBoon", "HeraSpecialBoon", "HeraCastBoon", "HeraSprintBoon" },
-			{ "HermesWeaponBoon", "HermesSpecialBoon", "TimedKillBuffBoon" },
-		},
-	},
-    boonIconPath = "Wistiti-HermesDuosBoonIcons\\HeraHermesDuo",
-	boonIconScale = 1.66,
-    
-	ExtractValues =
-	{
-		{
-			Key = "DamageShareAmountIncrease",
-			ExtractAs = "TooltipAmount",
-			Format = "Percent",
-		},
-		{
-			ExtractAs = "DamageShareDuration",
-			SkipAutoExtract = true,
-			External = true,
-			BaseType = "EffectData",
-			BaseName = "DamageShareEffect",
-			BaseProperty = "Duration",
-		},
-		{
-			ExtractAs = "DamageShareAmount",
-			SkipAutoExtract = true,
-			External = true,
-			BaseType = "EffectData",
-			BaseName = "DamageShareEffect",
-			BaseProperty = "Amount",
-			Format = "Percent",
-		},
-	},
-
-	ExtraFields = 
-	{
-		DamageShareAmountIncrease = { BaseValue = 0.2 },
-		SetupFunction = 
-		{
-			Name = "BuildValidEffects",
-			Args = 
-			{
-				StatusTraitNames = 
-				{
-					DamageEchoEffect = LinkedTraitData.ZeusEchoTraits,
-					BurnEffect = LinkedTraitData.HestiaBurnTraits,
-					ChillEffect = LinkedTraitData.DemeterRootTraits,
-					LegacyChillEffect = { "DemeterSprintBoon", "CastNovaBoon", "StormSpawnBoon" },
-					AmplifyKnockbackEffect = { "PoseidonStatusBoon", "PoseidonCastBoon" },
-					WeakEffect = LinkedTraitData.AphroditeWeakTraits,
-					BlindEffect = LinkedTraitData.ApolloBlindTraits,
-					DelayedKnockbackEffect = { "MassiveKnockupBoon" },
-					AresStatus = LinkedTraitData.AresRendTraits,
-				},
-				ConditionalOutgoingDamageMultipliers = 
-				{
-					{
-						EffectName = "AresStatus",
-						OutgoingDamageModifiers = 
-						{
-							ValidProjectilesLookup = { 
-								"HeraCastDamageProjectile",
-								"HeraSprintProjectile",
-							},
-							MissingEffectDamage = EffectData.AresStatus.BonusBaseDamageOnInflict,
-							MissingEffectName = "AresStatus",
-							MissingDamagePresentation = 
-							{
-								TextStartColor = Color.AresDamageLight,
-								TextColor = Color.AresDamage,
-								FunctionName = "AresRendApplyPresentation",
-								HitSimSlowParametersFalseTraitName = "StaffRaiseDeadAspect",
-								SimSlowDistanceThreshold = 180,
-								HitSimSlowCooldown = 0.8,
-								HitSimSlowParameters =
-								{
-									{ ScreenPreWait = 0.02, Fraction = 0.13, LerpTime = 0 },
-									{ ScreenPreWait = 0.10, Fraction = 1.0, LerpTime = 0.05 },
-								},
-							},
-						}
-					}
-				}
-			},
-		},
-		OnEffectApplyFunction = 
-		{
-			FunctionName = _PLUGIN.guid .. "." .. "HitchCopyStatus",
-			FunctionArgs = 
-			{
-				ValidStatusNames = 
-				{
-					DamageEchoEffect = true,
-					BurnEffect = "ApplyBurn",
-					ChillEffect = "ApplyRoot",
-					LegacyChillEffect = true,
-					AmplifyKnockbackEffect = true,
-					WeakEffect = "ApplyAphroditeVulnerability",
-					BlindEffect = true,
-					DelayedKnockbackEffect = true,
-					-- AresStatus = true,
-				},
-			},
-		},
-    },
-})]]
 
 -- Hermes x Poseidon
 gods.CreateBoon({
@@ -500,7 +289,7 @@ gods.CreateBoon({
 				ProjectileName = "DemeterOmegaStorm",
                 NumProjectiles = 2,
 				ProjectileCap = 3,
-				GustDamage = 7, --description only
+				GustDamage = 5, --description only
 				ReportValues = { 
 					ReportedGustDamage = "GustDamage",
 				},
@@ -596,7 +385,7 @@ gods.CreateBoon({
 	reuseBaseIcons = true,
 
     displayName = "Adoration Fee",
-    description = "Inflicting {$Keywords.Weak} on foes may temporarily {$Keywords.Charm} them, and their strikes gain you {#MoneyFormatBold}+5 {#Prev}{!Icons.Currency}.",
+    description = "Inflicting {$Keywords.Weak} may temporarily {$Keywords.Charm} most foes, and their strikes gain you {#MoneyFormatBold}+5 {#Prev}{!Icons.Currency}.",
 	StatLines = { "CharmChanceStatDisplay1" },
     customStatLine = {
         Id = "CharmChanceStatDisplay1",
@@ -710,7 +499,7 @@ gods.CreateBoon({
 	{
 		GoldtoArmorData = 
 		{
-			GoldCost = 10,
+			GoldCost = 5,
 			ArmorGain = 1,
 			ReportValues = { ReportedMultiplier = "GoldCost" },
 		},